refactor(fun): separate gambling service concerns
