+++ bot/features/fun/gambling/__init__.py
@@ -1,9 +1,15 @@
 """Interactive gambling workflow and Discord UI."""
 
+from .prompts import GamblingPromptMixin
 from .service import GamblingMixin
+from .session import GamblingSessionMixin
+from .settlement import GamblingSettlementMixin
 from .view import GambleView
 
 __all__ = [
     "GambleView",
     "GamblingMixin",
+    "GamblingPromptMixin",
+    "GamblingSessionMixin",
+    "GamblingSettlementMixin",
 ]
